History tab chart+date search + sku search
1.1 Violation trend by Incidents name will be Count of Violations by Sku

By default (when tab is just opened or show all clicked) it will show
result chart for whole history.
If product selected (or searched) it should show chart only for this
product
If Date selected it should show chart only for this Date
If Date Range and Product Selected it should show chart for this date
range and for this product.
If Single date selected (From and TO date is the same ) then
information for this Day.
If Dealer selected chart should show violations during selected date
range for this dealer, if date range not selected count of violation by
this dealer during all time